+++ changelog.md
@@ -1,2 +1,2 @@
-### Kube v0.3.2 release
-- fix input styling
+### Kube v0.3.3 release
+- Render utilities